Move claim_for_update to models.utils
